refactor: 파이차트 공통 컴포넌트화
