+++ src/api/dashboard/types.ts
@@ -1,0 +1,9 @@
+export type OverviewData = {
+  username: string;
+  totalValuation: number;
+  totalInvestment: number;
+  totalGain: number;
+  totalGainRate: number;
+  totalAnnualDividend: number;
+  totalAnnualDividendYield: number;
+};
